API

API documentation

Index

JuliaFEM.Elasticity

JuliaFEM.Element

JuliaFEM.Problem

JuliaFEM.update!

Types

JuliaFEM.Element — Type.

Construct a new element of type E.

Examples

julia> element = Element(Tri3, [1, 2, 3])

source

JuliaFEM.Problem — Type.

Construct a new field problem.

Examples

Create vector-valued (dim=3) elasticity problem:

julia> prob1 = Problem(Elasticity, "this is my problem", 3) julia> prob2 = Problem(Elasticity, 3)

source

Construct a new boundary problem.

Examples

Create Dirichlet boundary problem for vector-valued (dim=3) elasticity problem.

julia> bc1 = Problem(Dirichlet, "support", 3, "displacement") solver.

source

JuliaFEM.Elasticity — Type.

Elasticity equations.

Field equation is:

m∂²u/∂t² = ∇⋅σ - b

Weak form is: find u∈U such that ∀v in V

δW := ∫ρ₀∂²u/∂t²⋅δu dV₀ + ∫S:δE dV₀ - ∫b₀⋅δu dV₀ - ∫t₀⋅δu dA₀ = 0

where

ρ₀ = density
b₀ = displacement load
t₀ = displacement traction

Formulations

plane stress, plane strain, 3D

References

https://en.wikipedia.org/wiki/Linear_elasticity https://en.wikipedia.org/wiki/Finite_strain_theory https://en.wikipedia.org/wiki/Stress_measures https://en.wikipedia.org/wiki/Mooney%E2%80%93Rivlin_solid https://en.wikipedia.org/wiki/Strain_energy_density_function https://en.wikipedia.org/wiki/Plane_stress https://en.wikipedia.org/wiki/Hooke's_law

source

Functions

JuliaFEM.update! — Function.

Update time-dependent fields with new values.

Examples

julia> f = Field(0.0 => 1.0) julia> update!(f, 1.0 => 2.0)

Now field has two (time, value) pairs: (0.0, 1.0) and (1.0, 2.0)

Notes

Time vector is assumed to be ordered t_i-1 < t_i < t_i+1. If updating field with already existing time the old value is replaced with new one.

source

Update element field based on a dictionary of nodal data and connectivity information.

Examples

julia> data = Dict(1 => [0.0, 0.0], 2 => [1.0, 2.0]) julia> element = Seg2([1, 2]) julia> update!(element, "geometry", data)

As a result element now have time invariant (variable) vector field "geometry" with data ([0.0, 0.0], [1.0, 2.0]).

source

update!(problem.properties, attr...)

Update properties for a problem.

Example

update!(body.properties, "finite_strain" => "false")

source

Update problem solution vector for assembly.

source

Update solution from assebly to elements.

source

Default update for solver.

source