API documentation

API documentation

Index

FEMBase.AbstractField

FEMBase.CVTV

FEMBase.DCTI

FEMBase.DCTV

FEMBase.DVTI

FEMBase.DVTId

FEMBase.DVTV

FEMBase.DVTVd

FEMBase.Element

FEMBase.Problem

JuliaFEM.Elasticity

FEMBase.assemble_elements!

FEMBase.field

FEMBase.update!

FEMBasis.interpolate

FEMBasis.interpolate

Types

FEMBase.Element — Type.

Element(topology, connectivity)

Construct a new element where topology is the topological type of the element and connectivity contains node numbers where element is connected.

Topological types

1d elements

Seg2

Seg3

2d elements

Tri3

Tri6

Tri7

Quad4

Quad8

Quad9

3d elements

Tet4

Tet10

Hex8

Hex20

Hex27

Pyr5

Wedge6

Wedge15

Examples

element = Element(Tri3, (1, 2, 3))

FEMBase.Problem — Type.

Problem{P<:AbstractProblem}

Defines a new problem of type P, where P characterizes the physics of the problem. P can be for example Elasticity, if the physics of the system is described by Cauchy's stress equation ∇⋅σ + b = ̈ρu, or Heat, if the physics of the problem is described by heat equation -∇⋅(k∇u) = f.

JuliaFEM.Elasticity — Type.

Elasticity equations.

Field equation is:

m∂²u/∂t² = ∇⋅σ - b

Weak form is: find u∈U such that ∀v in V

δW := ∫ρ₀∂²u/∂t²⋅δu dV₀ + ∫S:δE dV₀ - ∫b₀⋅δu dV₀ - ∫t₀⋅δu dA₀ = 0

where

ρ₀ = density
b₀ = displacement load
t₀ = displacement traction

Formulations

plane stress, plane strain, 3D

References

https://en.wikipedia.org/wiki/Linearelasticity https://en.wikipedia.org/wiki/Finitestraintheory https://en.wikipedia.org/wiki/Stressmeasures https://en.wikipedia.org/wiki/Mooney%E2%80%93Rivlinsolid https://en.wikipedia.org/wiki/Strainenergydensityfunction https://en.wikipedia.org/wiki/Planestress https://en.wikipedia.org/wiki/Hooke'slaw

source

Functions

FEMBase.update! — Function.

update!(elements, field_name, data)

Given a list of elements, field name and data, update field to elements. Data is passed directly to the field-function.

Examples

Create two elements with topology Seg2, one is connecting to nodes (1, 2) and the other is connecting to (2, 3). Some examples of updating fields:

elements = [Element(Seg2, [1, 2]), Element(Seg2, [2, 3])]
X = Dict(1 => 0.0, 2 => 1.0, 3 => 2.0)
u = Dict(1 => 0.0, 2 => 0.0, 3 => 0.0)
update!(elements, "geometry", X)
update!(elements, "displacement", 0.0 => u)
update!(elements, "youngs modulus", 210.0e9)
update!(elements, "time-dependent force", 0.0 => 0.0)
update!(elements, "time-dependent force", 1.0 => 100.0)

When using dictionaries in definition of fields, key of dictionary corresponds to node id, that is, updating field geometry in the example above is updating values (0.0, 1.0) for the first elements and values (1.0, 2.0) to the second element. For time dependent field, syntax time => data is used. If field is initialized without time-dependency, it cannot be changed to be time-dependent afterwards. If unsure, it's better to initialize field with time dependency.